requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.0
chromadb>=0.4.0
google-generativeai>=0.8.0
sentence-transformers>=2.7.0
